move function used in tests only
