+++ core/relay.go
@@ -41,10 +41,10 @@ func NewRelay(ethClient *ethclient.Client, config *config.Config) *Relay {
 
 	balanceOfBonded := func(account common.Address) (*big.Int, *relayerror.Error) {
 		balance, err := atlETHContract.BalanceOfBonded(nil, account)
-
 		if err != nil {
 			return nil, ErrCantGetBondedBalance.AddError(err)
 		}
+
 		return balance, nil
 	}
 
@@ -88,6 +88,7 @@ func (r *Relay) submitUserOperation(userOp *operation.UserOperation) (common.Has
 	if relayErr != nil {
 		return common.Hash{}, relayErr
 	}
+
 	go r.server.BroadcastUserOperation(userOp)
 	return userOpHash, nil
 }
